#ENH improved support for GE P-files
The previous approach to supporting GE P-files was to have users
build a derived Docker image which would both set up the necessary
foundations for Orchestra to run, then copy in a local version of the
SDK (which cannot be distributed with OpenMRSLab). This was a bit clunky
so now we include the foundation steps (building an updated version of
OpenSSL and renaming some of the libraries) in stock OpenMRSLab, as this
does not break anything else, and adding a hypothetical "orchestra"
folder inside the Jupyter work directory to the PythonPath. Thus all the
user needs to do now is create that folder, either through a mapped
folder or uploading via Jupyter itself, and copy the SDK into it.

We have also taken the opportunity to update the Jupyter base container
which is now running on Ubuntu 20.04, and Python 3.8. GCC has been
downgraded from the default of 9 back to 8 as there are some issues
compiling NiftyReg with version 9 (which relate to OpenMP in some way).

We also upgrade to Suspect v0.4.3.